requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
openai==2.8.1
python-dotenv==1.0.1
pydantic==2.8.2
